+++ src/concurrency.ts
@@ -32,15 +32,19 @@ export class Concurrency extends SharedBase<ConcurrencyCommonOptions> {
         const iterator = validateAndProcessInput(input);
         let done = false;
 
+        const catchAndAbort = (err: unknown) => {
+            done = true;
+            throw err;
+        };
+
         await new Promise<void>((resolve, reject) => {
+            let count = maxConcurrency;
             for (let i = 0; i < maxConcurrency; i++) {
                 (async () => {
                     while (!done) {
                         const data = await iterator.next();
-                        if (done || data.done) {
-                            done = true;
-                            return;
-                        }
+                        if (data.done || done)
+                            break;
 
                         const result = await task(await data.value);
                         if (result === interrupt) {
@@ -54,7 +58,8 @@ export class Concurrency extends SharedBase<ConcurrencyCommonOptions> {
                         }
                     }
                 })()
-                    .then(() => (--i === 0) ? resolve() : undefined)
+                    .then(() => (--count === 0) ? resolve() : undefined)
+                    .catch(catchAndAbort)
                     .catch(reject);
             }
         });
@@ -273,12 +278,13 @@ export class Concurrency extends SharedBase<ConcurrencyCommonOptions> {
         let done = false;
         const { maxConcurrency } = this.#options;
 
-        const catchAndAbort = (err: any) => {
+        const catchAndAbort = (err: unknown) => {
             done = true;
             throw err;
         };
 
         await new Promise<void>((resolve, reject) => {
+            let count = maxConcurrency;
             for (let i = 0; i < maxConcurrency; i++) {
                 this.run(async () => {
                     while (true) {
@@ -295,7 +301,7 @@ export class Concurrency extends SharedBase<ConcurrencyCommonOptions> {
                         })();
                     }
                 })
-                    .then(() => (--i === 0) ? resolve() : undefined)
+                    .then(() => (--count === 0) ? resolve() : undefined)
                     .catch(catchAndAbort)
                     .catch(reject);
             }