+++ tests/concurrency.test.ts
@@ -8,7 +8,7 @@ async function wait(ms: number) {
     return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
-test('Iterable', async t => {
+test('forEach', async t => {
     function* test() {
         yield 1;
         yield 2;
@@ -28,6 +28,51 @@ test('Iterable', async t => {
     });
 
     t.deepEqual(calls, [1, 2, 3, 4]);
+    t.pass();
+});
+
+test('map', async t => {
+    function* test() {
+        yield 1;
+        yield 2;
+        yield 3;
+        yield 4;
+        return;
+    }
+
+    const calls = await Concurrency.map({
+        input: test(),
+        maxConcurrency: MAX_CONCURRENCY,
+        task: async (value) => {
+            await wait(value * 10);
+            return value;
+        }
+    });
+
+    t.deepEqual(calls, [1, 2, 3, 4]);
+    t.pass();
+});
+
+test('mapSettled', async t => {
+    function* test() {
+        yield 1;
+        yield 2;
+        yield 3;
+        yield 4;
+        return;
+    }
+
+    const calls = await Concurrency.mapSettled({
+        input: test(),
+        maxConcurrency: MAX_CONCURRENCY,
+        task: async (value) => {
+            await wait(value * 10);
+            return value;
+        }
+    });
+
+    const result = calls.filter(x => x.status === 'fulfilled').map(x => x.value);
+    t.deepEqual(result, [1, 2, 3, 4]);
     t.pass();
 });